+++ src/game/continuous_effects.rs
@@ -748,6 +748,7 @@ impl Game {
             | EffectRecipientDef::ObjectsSharingNameWithTarget(_)
             | EffectRecipientDef::TriggeringObject
             | EffectRecipientDef::ControllerOfTriggeringObject
+            | EffectRecipientDef::ControllerOfAttachedPermanent
             | EffectRecipientDef::EventPlayer => false,
         }
     }